Modify all wrappers to use bash
This patch fixes the problem caused by #11 that the backslash
replacements in msvcenv.sh is not supported by /bin/sh. Considering that
the nmake wrapper has already been using bash, we changed all wrappers to use
bash for consistency.

This patch is not yet tested from a clean build because of #12.