vsdownload: Add an option for preferring a different SDK version
Normally, we do get a specific version as a recommended dependency,
so we need to explicitly ignore that. For installing specific, older
versions, it would be easier to override what version to install,
but this approach (ignoring all other versions but the chosen one)
works both for the default (latest) and the older specific versions.