Allow to use vsdownload on Windows
The script can actually be useful to download VS on Windows too. When
doing so, use the Windows-native msiexec tool to extract the MSIs,
instead of msiextract.

And because of how open files are handled on Windows, use a context
manager for the subprocess.check_call output, so that if it throws an
exception, the file is not left open.