test: Test building Meson based projects with msvc-wine
Only testing building with MSVC - not testing building with
clang-cl with Meson. Clang-cl does work with Meson, but for the
resource compiler, Meson currently fails to detect llvm-rc.

Therefore, currently if building with Meson, one has to use the
Wine wrapped version of rc.exe, which requires Wine to be available,
contrary to the rest of the clang-cl/lld based setup. (The CI
environment for the clang/lld tests intentionally does not have Wine
available.)