wrappers: Disable STDOUT/STDERR filtering for MSBuild
MSBuild.exe produces nice colored output in the terminal when
its STDOUT and STDERR are attached directly to the TTY.
The command line arguments are still processed as usual.

Refactor wine-msvc.sh a little and make it honor the WINEDEBUG
environment variable value set by the users to aid debugging.