vsdownload: Only download the right SDK for MSVC 16/17 with --msvc-version
Previously, we'd end up downloading both the default SDK from the
manifest, and the SDK that was default for that version of MSVC.

Then install.sh would pick the latest SDK out of the ones installed,
which wouldn't be the right choice when we explicitly wanted an older
one.

If selecting an implicit default SDK via --msvc-version, set this
as the option as if set via --sdk-version, which forces other
SDKs to be ignored.