github: Test with a different versions of old toolchains
The manifest for --major 15 no longer includes older WinSDK versions.
When attempting to install older toolchains, the install ends up
without any WinSDK at all. But when attempting to install 15.9
it does end up pulling in WinSDK 10.0.19041, so the install is
working, even if it isn't the exact SDK that we did try to install.

For --major 16, it clearly errors out saying "WinSDK version
10.0.17763 not found", and listing that 10.0.19041 and
10.0.22621 are available. Test installing 16.9 which defaults
to WinSDK 10.0.19041.